refactor(tasks): migrate preprocess module Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
