+++ src/pyrovelocity/tasks/__init__.py
@@ -1,6 +1,7 @@
 import pyrovelocity.tasks.data
-
+import pyrovelocity.tasks.preprocess
 
 __all__ = [
     "data",
+    "preprocess",
 ]
