+++ src/pyrovelocity/tests/test_preprocess.py
@@ -1,7 +1,7 @@
-"""Tests for `pyrovelocity.preprocess` module."""
+"""Tests for `pyrovelocity.tasks.preprocess` module."""
 
 
 def test_load_preprocess():
-    from pyrovelocity import preprocess
+    from pyrovelocity.tasks import preprocess
 
     print(preprocess.__file__)
